Split classes from extract.py and added parameters to the config.
