+++ classifier_TEMPLATE.yaml
@@ -1,0 +1,113 @@
+
+base_data_folder: <put your base folder here>
+
+# source_folder is where the cptv files reside under (could be in a subdirectory). It is specified relative
+# to the base_data_folder
+source_folder: "cptv"
+
+# which folders (tags) to ignore when selecting cptv files from source_folder
+excluded_folders: ['other', 'unidentified', 'untagged', 'moving', 'multi','missclassified']
+
+# tracks_folder is where the track files will be created.  It is specified relative to the base_data_folder
+tracks_folder: "tracks"
+
+# Overwrite existing files.  Options are all, old, or none.'
+overwrite_mode: "old"
+
+# colour map to use when exporting to MPEG.
+# Note:  This is relative to the current directory of the application, not the base data folder
+previews_colour_map: "custom_colormap.dat"
+
+tracking:
+    #
+    #  Tracking algorithm
+    #
+
+    # auto threshold needs to find a near maximum value to calculate the threshold level
+    # a better solution might be the mean of the max of each frame?
+    threshold_percentile: 99.9
+
+    # measures how different pixels are from estimated background, if they are on average less different than this then
+    # video is considered to be static.
+    static_background_threshold: 4.0
+
+    # any clips with a mean temperature hotter than this will be excluded
+    max_mean_temperature_threshold: 10000
+
+    # any clips with a temperature dynamic range greater than this will be excluded
+    max_temperature_range_threshold: 10000
+
+    # number of pixels around object to pad.
+    frame_padding: 6
+
+    # number of frames to wait before removing a track from the list of tracks
+    remove_track_after_frames: 9
+
+    # strategy to use when dealing with regions of interest that are cropped against the side of the frame
+    # in general these regions often do not have enough information to accurately identify the animal.
+    # options are
+    # 'all': All cropped regions are included, good for classifier
+    # 'cautious': Regions that are only cropped a bit are let through, this is good for training data
+    # 'none': No cropped regions are permitted.  This is the most safe.
+    cropped_regions_strategy: "cautious"
+
+    # when enabled smooths tracks so that track dimensions do not change too quickly.
+    track_smoothing: False
+
+    high_quality_optical_flow: False
+
+    # minimum allowed threshold for number of pixels changed from average background, smaller values detect more objects, but bring up additional false positives
+    min_threshold: 30
+
+    # minimum allowed threshold for number of pixels changed from average background, smaller values detect more objects, but bring up additional false positives
+    max_threshold: 50
+
+    # how much to threshold thermal before calculating optical flow.
+    flow_threshold: 40
+
+    # maximum number of tracks to extract from a clip.  Takes the n best tracks.
+    max_tracks: 10
+
+    filters:
+        # discard any tracks that overlap with other tracks more than this.   T than this length in seconds
+        track_overlap_ratio: 0.5
+
+        # discard any tracks shorter than this length in seconds
+        min_duration_secs: 3.0
+
+        # dicard any tracks that do not move enough, (move less than this)
+        track_min_offset: 4.0
+
+        # discard tracks that do not have enough delta within the window (i.e. pixels that change a lot)
+        track_min_delta: 1.0
+
+        # discard tracks that do not have enough enough average mass.
+        track_min_mass: 2.0
+
+
+
+
+
+    #
+    #  Tracking processor
+    #
+
+    # enables fast BLOSC compression (requires plugin)
+    enable_compression: False
+
+    # includes the filtered channel when exporting.  This is typically not used.  If compression is enabled the
+    # filesize can be reduced by not including it.
+    include_filtered_channel: False
+
+    # Number of worker threads to use.  0 disables worker pool and forces a single thread.
+    worker_threads: 0
+
+    # Create a preview of each track
+    preview_tracks: False
+
+    # Hints file to loade.  Hints are a way to give extra information to the tracker about how many tracks to take.
+    # Note:  Set to 'False' is there is no hints file.
+    # Note:  This is relative to the current directory of the application, not the base data folder
+    hints_file: "hints.txt"
+
+
