+++ ml_tools/tools.py
@@ -203,9 +203,9 @@ def stream_mpeg(filename, frame_generator):
 
         return_code = process.wait(timeout=30)
         if return_code != 0:
-            raise Exception("FFMPEG failed with error {}".format(return_code))
+            raise Exception("FFMPEG failed with error {}. Have you isntalled ffmpeg and added it to your path?".format(return_code))
     except Exception as e:
-        logging.error("Failed to write MPEG: %s", e)
+        logging.error("Failed to write MPEG: %s.  Have you isntalled ffmpeg and added it to your path?", e)
         if process is not None:
             logging.error(process.stderr.read())
 
@@ -327,7 +327,7 @@ def get_session(disable_gpu=False):
 
     global tf
     import tensorflow as tf
-    
+
     session = None
 
     if disable_gpu:
